after groups duplicated code cleanup
